refactor: 코드 하이라이팅 로직 훅 분리
